% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fft_params.R
\name{get_fft_params}
\alias{get_fft_params}
\title{Calculate FFT Parameters}
\usage{
get_fft_params(
  samp.rate = NULL,
  win.length = NULL,
  time.res = NULL,
  freq.res = NULL
)
}
\arguments{
\item{samp.rate}{Numeric. The sampling rate of the audio signal in Hz.}

\item{win.length}{Numeric. The window length in samples. Can be `NULL` if time or frequency resolution is provided.}

\item{time.res}{Numeric. The time resolution in seconds. Can be `NULL` if window length or frequency resolution is provided.}

\item{freq.res}{Numeric. The frequency resolution in Hz. Can be `NULL` if window length or time resolution is provided.}
}
\value{
Prints the calculated time and frequency resolution, or window length and either time or frequency resolution depending on the provided inputs.
}
\description{
This function calculates the time and frequency resolution based on the provided sampling rate and either window length, time resolution, or frequency resolution. Depending on which parameters are given, it calculates and returns the appropriate values for the others.
}
\examples{
get_fft_params(samp.rate = 44100, win.length = 1024)
get_fft_params(samp.rate = 44100, time.res = 0.023)
get_fft_params(samp.rate = 44100, freq.res = 43)
}
