% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_index.R
\name{harmonize_index}
\alias{harmonize_index}
\title{Harmonize "old" soundecology2 output with new SoundEcology2 table format}
\usage{
harmonize_index(data.frame, original.index = NULL)
}
\arguments{
\item{data.frame}{A data frame containing the results from soundecology2}

\item{original.index}{Character. A name for the 'old' index, e.g., "adi_original"}
}
\value{
A new data frame with the harmonized columns.
}
\description{
Harmonize "old" soundecology2 output with new SoundEcology2 table format
}
\examples{
harmonize_index(ndsi_original, "ndsi_original")
}
