% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_se2_indices.R
\name{merge_se2_indices}
\alias{merge_se2_indices}
\title{Merge CSV outputs of several indices calculated with SoundEcology2}
\usage{
merge_se2_indices(folder_path = NULL, recursive = TRUE)
}
\arguments{
\item{folder_path}{character. The path to the folder containing the csv files with the "results.csv" suffix}

\item{recursive}{logical. If TRUE, it looks for csv files in the subfolders.}
}
\value{
A tibble with the merged tables for different indices.
}
\description{
Merge CSV outputs of several indices calculated with SoundEcology2
}
\examples{
\dontrun{
indices_all <- merge_se2_indices("path_to_folder", recursive = TRUE)
}
}
